REF: Move filestore and portable-fs into resource_registry.
