+++ databroker/resource_registry/path_only_handlers.py
