+++ databroker/resource_registry/tests/test_commands.py
