+++ databroker/resource_registry/tests/test_mongo_details.py
